Fix type error from new bindgen configuration

+++ src/core/ruby_version.rs
@@ -809,7 +809,7 @@ macro_rules! get_cfunc_name(
                 imemo_type_imemo_ment => Ok(&imemo as *const rb_method_entry_struct),
                 imemo_type_imemo_svar => {
                     let svar: vm_svar = source.copy_struct(raw_imemo).context(raw_imemo)?;
-                    check_method_entry(svar.cref_or_me, source)
+                    check_method_entry(svar.cref_or_me as usize, source)
                 },
                 _ => Ok(raw_imemo as *const rb_method_entry_struct)
             }
